CHANGE THE WAY MESSAGE TRACING IS DONE
This change is necessary because currently messages are implemented such
that their record of senders includes the object that created the
message even before it is sent.

    EX.
       * Assume there are objects 1-3 that can send messages.
       * send(1, 2) means that a message is sent from object 1 to object 2
       * "process" is a list of sequential events
       * "source" is the property of the message object logging all
         objects that have sent it. It corresponds to the process above
         it.

       process: creation by 1 -> send message from 1 to 2 -> send from 2 to 3
        source:     [1] ------->         [1, 1] ----------->     [1, 1, 2]

The changes no longer provide a means for specifying the initial
contents of a Message's source property. Instead, a method is provided
on the Message object which can be used to append a sender to the source
property and which returns the Message afterword so that Message sending
can be accomplished in a single line.