MAKE ARCH_ELEMENT EXTEND THREAD INSTEAD OF PROCESS
This change is necessary so that the message passing activities of
arch_elements do not block as a process. This may have the potential to
cause deadlock. Furthermore, the context switching for processes is done
at the operating system level so, if a component only forwards messages
(like basic connectors do), then it never 'needs' a separate process
since it is basically IO bound.