ADD 'CONDITIONAL IMMUTABILITY' TO EVENT
This change is necessary because Event copying, especially for Events
with a large payload is the most expensive operation in the framework.
To reduce the computational load of Events copying, Events are now
passed between components by reference instead of by value. This
ensures that copying will only occur in places where copying must
*necessarily* occur.

As documented in the event.py file, this has some interesting
implications. However, everything will work as it should if components
simply refrain from mutating Events unless there exists exactly one
reference to that message in the system. At the moment, I believe that
the only Events that meet this requirement are those that have just been
created and have not yet been sent. If the contents of another message
must be mutated, a component must mutate a copy instead.