ADD THREAD MANIPULATION OPERATIONS TO ARCH_ELEM
ArchElements now have a default run() method which calls their
behavior() method in a continuous loop. This is done in such a way that
an ArchElement will be able to listen for and react to incoming messages
from an ArchitectureMessenger in the future. Based on these messages, an
ArchElement may be paused, resumed, or stopped.