ADD DEFAULT BEHAVIOR TO COMPONENT
As the title says. Additionally, the component will default to it's
behavior() method at initialization if the constructor is not passed a
behavior callable. This means that the default behavior of a component
can be overridden after it is initialized (or even changed while it is
running!) simply by reassigning its behavior attribute to a new
callable.